Updated ComposeProcessor to new API
