+++ streampipes-processors-filters-jvm/src/main/java/org/apache/streampipes/processors/filters/jvm/FiltersJvmInit.java
@@ -29,7 +29,7 @@ import org.apache.streampipes.messaging.jms.SpJmsProtocolFactory;
 import org.apache.streampipes.messaging.kafka.SpKafkaProtocolFactory;
 import org.apache.streampipes.messaging.mqtt.SpMqttProtocolFactory;
 import org.apache.streampipes.processors.filters.jvm.processor.booleanfilter.BooleanFilterProcessor;
-import org.apache.streampipes.processors.filters.jvm.processor.compose.ComposeController;
+import org.apache.streampipes.processors.filters.jvm.processor.compose.ComposeProcessor;
 import org.apache.streampipes.processors.filters.jvm.processor.enrich.MergeByEnrichProcessor;
 import org.apache.streampipes.processors.filters.jvm.processor.limit.RateLimitController;
 import org.apache.streampipes.processors.filters.jvm.processor.merge.MergeByTimeProcessor;
@@ -61,7 +61,7 @@ public class FiltersJvmInit extends StandaloneModelSubmitter {
                     new MergeByEnrichProcessor(),
                     new MergeByTimeProcessor(),
                     new MergeBySchemaProcessor(),
-                    new ComposeController(),
+                    new ComposeProcessor(),
                     new NumericalTextFilterProcessor(),
                     new RateLimitController())
             .registerMessagingFormats(
